Do not truncate variability (21911398)
SPRT(-5.66, 0), tc=8+0.08, hash=8:

time vs master: 967 - 904 - 1780  [0.509] 3651
Elo difference: 6.00 +/- 8.06
SPRT: llr 2.95, lbound -2.94, ubound 2.94 - H1 was accepted

+++ src/uci.h
@@ -22,4 +22,4 @@ void info_update(Info *info, int depth, int score, int64_t nodes, move_t pv[], b
 void info_print_bestmove(Info *info);
 move_t info_best(Info *info);
 int info_last_depth(Info *info);
-int info_variability(Info *info);
+double info_variability(Info *info);
